Rename `vertex.secrets` ~> `vertex.vertex_secrets` and update all refs (addresses issue #66)
